chg: modify mask structure as a folder. New: add three new masks
