--- pygaming/screen/mask.py
@@ -1,84 +1,0 @@
-from abc import ABC, abstractmethod
-import numpy as np
-from pygame import Surface, surfarray as sa
-from ..error import PygamingException
-import cv2
-
-# Mask effects
-ALPHA = 'alpha'
-DARKEN = 'darken'
-LIGHTEN = 'lighten'
-SATURATE = 'saturate'
-DESATURATE = 'desaturate'
-
-_EFFECT_LIST = [ALPHA, DARKEN, LIGHTEN, SATURATE, DESATURATE]
-
-class Mask(ABC):
-    """Mask is an abstract class for all masks."""
-    
-    def __init__(self, width: int, height: int) -> None:
-        super().__init__()
-        self._loaded = False
-        self._width = width
-        self._height = height
-        self.matrix = None
-
-    @abstractmethod
-    def _load(self):
-        raise NotImplementedError()
-    
-    def load(self):
-        self._load()
-        self._loaded = True
-
-    @abstractmethod
-    def unload(self):
-        self.matrix = None
-        self._loaded = False
-    
-    def apply(self, surface: Surface, effects: dict[str, float]):
-        if not self._loaded:
-            self.load()
-        
-        if surface.get_size() != (self._width, self._height):
-            raise PygamingException("The size of the mask do not match the size of the art.")
-
-        if ALPHA in effects:
-            surf_alpha = sa.array_alpha(surface)
-            surf_alpha[:] = np.astype(np.clip(surf_alpha * self.matrix * effects[ALPHA], 0, 255), surf_alpha.dtype)
-
-        if any(effect in _EFFECT_LIST for effect in effects):
-            rgb_array = sa.pixels3d(surface)
-            hls_array = cv2.cvtColor(rgb_array, cv2.COLOR_RGB2HLS)
-
-            if DARKEN in effects:
-                hls_array[:,:, 1] = hls_array[:,:, 1] * (1 - self.matrix * effects[DARKEN]) 
-            
-            if LIGHTEN in effects:
-                hls_array[:,:, 1] = 255 - (255 - hls_array[:,:, 1]) * (1 - self.matrix * effects[LIGHTEN]) 
-
-            if DESATURATE in effects:
-                hls_array[:,:, 2] = hls_array[:,:, 2] * (1 - self.matrix * effects[DESATURATE])
-            
-            if SATURATE in effects:
-                hls_array[:,:, 2] = 255 - (255 - hls_array[:,:, 2]) * (1 - self.matrix * effects[SATURATE])
-
-            rgb_array[:] = cv2.cvtColor(hls_array, cv2.COLOR_HLS2RGB)[:]
-    
-    def __bool__(self):
-        return True
-
-
-class MatrixMask(Mask):
-    """A matrix mask is a mask based on a matrix."""
-    
-    def __init__(self, width: int, height: int, matrix: np.ndarray = None) -> None:
-        super().__init__(width, height)
-        self.matrix = np.clip(matrix, 0, 1)
-
-    def unload(self):
-        """Don't do anything as we want to keep the matrix."""
-    
-    def _load(self):
-        """Don't do anything as the matrix is already loaded."""
-
